coffee_stained_map/requirements.txt
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==4.2.14
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==23.5
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==37.12.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-decouple==3.8
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
